many small changes, most important is maybe the adding of new instructions.
limit the length of the written offspring genome.
change instructions "M0", "L0" to "MRK0", "LBL0".
make the world smaller in y direction (to be able to see the log messages in a terminal).
change clouds: they don't drop many drops of 1 energy, but a configurable drop size and a correspondingly smaller number of drops.
change genome of the starting organism to accomodate the new instructions.
raise an error if the starting organism doesn't compile.
upon calling (default-world), reset the random state.
upon cell division the parent cell doesn't die anymore and only one daughter cell is created.
fix a bug that occurred when an organism was waiting for more than one loop: it was killed.
do not make random jumps to other instructions as that would be taxing to implement in assembler.
change waiting times for eating and walking.
add instructions for adding registers an, bn, for subtraction and multiplication.
make cell division not move the daughter cell into a direction, but let it stay at the same coordinate as the mother cell.
add walking into y direction.
change default number of instructions per loop and per organism from 30 to 200.
change maximal age of an organism without cell division from 100 iterations to 1000.
make random color generation for a selected organism use a different random state to not influence the randomness in evolution by user interactions.
when upon starting (software-render-texture) there are no organisms, start with the default organisms.
pressing key "o" changes the status printing to overview and clears the cursor selection.